many changes to RL Matrix
A property for keeping track of the agent
Changed the memory replay to a ```queue``` in order to fix capacity that is being ignored (ie, memory leak)
Changed ```Step()``` to return ```Task<int>``` instead of ```Task``` so that I could keep track of the selected action
In the example environments, the ```GetCurrentState()``` reset the environment. I only wanted to reset the environment at the end of the step. Therefore I updated the ```Step()``` so that I could get the pre and post state of the environment without resetting it on the first ```GetStateAsync()```
Finally, the ability to pass in ActionsMask(int[] mask)